requirements.txt
# requirements.txt (обновлённый)
fastapi==0.111.0
uvicorn[standard]==0.25.0
pydantic==2.5.1
redis==5.3.1
torch
torchvision
numpy
pillow
albumentations